Perform merge with --allow-unrelated-histories
With our shallow clones, fetching only the tips of `main` and `rc`, we
were not able to actually perform a merge, receiving the following
error:

    fatal: refusing to merge unrelated histories

Based on the contents of the local checkout, these branches did really
appear to be unrelated, but only because we didn't have the entire
history; if we were to do a full clone, they would of course be
related.

Rather than trying to fetch code until we get to the most recent
common ancestor of both branches (or do a full fetch, which rather
defeats the point of shallow clones in the first place), we can
instead just tell `git` to ignore this situation and perform a merge
anyway.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>