Ensure we're operating on "real" (not grafted) commits
After rolling out the ability to work on shallow clones, we noticed
some problems generating the merge commit from `main` to `rc`, in that
certain changes from the `main` branch were being excluded from the
merge.

This appears to have been because our `rc` commit, being the tip of a
`depth=1` shallowly-fetched branch, was what is known as a "grafted"
commit, which means it effectively contains the entire history of that
branch in a single commit. Because of how this interacted with our
merging strategy and our `--allow-unrelated-histories` option, we
ended up privigeling *too much* content on the `rc` branch, instead of
from the `main` branch during the merge.

Here, we switch to a `depth=2` fetch, which ensures that the tip is
actually a real commit, and does not contain the entire history.

We *will* need to make some additional commits to the various `rc`
branches that have been affected by this code, to ensure they pick up
all the code changes they need. Once that is done, however, the
current commit should ensure that things work properly going forward.

As a further guard, we now explicitly test to make sure the tip of the
`rc` branch is not a grafted commit.

Further detail is contained in code comments.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>