Revise RC logic: use "theirs" strategy, and lose the mocks
Previously, we were using an "ours" merge strategy option as a way to
preserve the version pins from the `rc` branch. While this worked
initially (since the only possible difference between `rc` and `main`
would be the version pins), it has since been shown to be the wrong
way to go about things. For various reasons, we have had to make
commits directly to the `rc` branch to deal with some issues that
arose from the use of shallow clones. With the "ours" strategy, this
ends up preserving these changes on the `rc` branch, regardless of
which changes have happened on `main`.

As a result, we now switch to the "theirs" strategy option, which
retains everything from the `main` branch. As a happy side effect,
this also means we can simplify our version pinning resolution
logic. Rather than doing a complicated saving of `rc` branch version
pins, capturing of the `main` branch configuration files, layering the
`rc` pins on top, and then layering the new version pins, we can rely
on the merge to give us our `main` configuration layer, and then just
extract the pins from the rc branch using `git show`. This is slightly
more straightforward, and easier to reason about.

To ensure that we are properly dealing with the commits in the right
way (as well as making sure we handle shallow clones), we now create a
small testing Git repository within the context of the integration
tests. Rather than mocking `git`, we allow the plugin to actually
perform the real `git` operations on a real repository. We also
perform some basic sanity tests to verify that the checkout we're
operating on is, in fact, shallow, and that merge logic proceeds as we
expect.

To aid in this, we have to modify the testing container a bit to
include `git`, as well as `yq` (think `jq`, but for YAML) to make our
Pulumi mocks more true-to-life (and compatible with running actual
`git` commands).

Signed-off-by: Christopher Maier <chris@graplsecurity.com>