Add a `log_and_run` helper
To help get as much visibility as possible into what is happening, we
run all `pulumi` and `git` invocations through a new `log_and_run`
helper function.

This function prints the command that is being invoked (nicely
formatted with ANSI codes), and then runs it. As such, it is just a
simple adapter that adds logging, but doesn't affect anything else.

(The logging and ANSI code is structured as it is because I basically
swiped it from another thing I've written that had more ANSI and more
logging functions. Those can come later if we want / need them.)

Signed-off-by: Christopher Maier <chris@graplsecurity.com>