Update `buildkite/plugin-tester` image to v2.0.0
There's finally a formal release of `buildkite/plugin-tester` that we
can use! This locks us to that new release, and addresses a breaking
change that was introduced.

See
https://github.com/buildkite-plugins/buildkite-plugin-tester/releases/tag/v2.0.0

The version of Alpine the new image is based on has been updated,
which means that we no longer have to add the `edge` APK repository to
get a version of `git` that we can use. While we were using `git`
2.36.1, we're now using 2.34.2; this is just because that's what's in
the standard APK repository for this Alpine release. As long as it's
at least 2.33, our tests will be fine.

Also adds some tweaks suggested by Hadolint, now that that is wired
up.

Signed-off-by: Christopher Maier <chris@graplsecurity.com>